+++ src/Components/Loader/Loader.css
@@ -25,7 +25,7 @@
     left: 0;
     width: 100%;
     height: 100%;
-    background-color: black;
+    background-color: #000000;
     opacity: 0.4;
 }
 
@@ -43,10 +43,15 @@
 /* Spinner Circle Rotation */
 .sp-circle {
     border: 4px rgba(0, 0, 0, 0.25) solid;
-    border-top: 4px black solid;
+    border-top: 4px #000000 solid;
     border-radius: 50%;
     -webkit-animation: spCircRot 0.6s infinite linear;
     animation: spCircRot 0.6s infinite linear;
+}
+
+.loader--dark .sp-circle{
+    border: 4px rgba(255, 255, 255, 0.5) solid;
+    border-top: 4px #ffffff solid;
 }
     
 @-webkit-keyframes spCircRot {
