+++ src/Components/index.js
@@ -1,2 +1,5 @@
 export { Footer } from './Footer/Footer';
 export { Navbar } from './Navbar/Navbar';
+export { CategoryCard } from './CategoryCard/CategoryCard';
+export { Loader } from './Loader/Loader'
+export { LoaderOverlay } from './Loader/LoaderOverlay';
