+++ src/Context/index.js
@@ -1,0 +1,2 @@
+export * from './AuthContext/AuthContext';
+export * from './ThemeContext/ThemeContext';
